rename structure to verifiers
