+++ src/rules/ts-config-esmoduleinterop.ts
@@ -5,7 +5,7 @@
 
 "use strict";
 
-import structure from "../utils/structure";
+import getVerifiers from "../utils/verifiers";
 import { Rule } from "eslint";
 
 //------------------------------------------------------------------------------
@@ -27,7 +27,7 @@ export = {
     schema: [] // no options
   },
   create: (context: Rule.RuleContext): Rule.RuleListener => {
-    const verifiers = structure(context, {
+    const verifiers = getVerifiers(context, {
       outer: "compilerOptions",
       inner: "esModuleInterop",
       expected: true,
