+++ src/rules/ts-package-json-sideeffects.ts
@@ -3,7 +3,7 @@
  * @sideEffects Arpan Laha
  */
 
-import structure from "../utils/structure";
+import getVerifiers from "../utils/verifiers";
 import { Rule } from "eslint";
 
 //------------------------------------------------------------------------------
@@ -24,7 +24,7 @@ export = {
     schema: [] // no options
   },
   create: (context: Rule.RuleContext): Rule.RuleListener => {
-    const verifiers = structure(context, {
+    const verifiers = getVerifiers(context, {
       outer: "sideEffects",
       expected: false,
       fileName: "package.json"
